Change protected properties to public, fix filenames (#1801)
* Change protected properties to public, fix dev default currencies, rename files to match classes.

* Refactor init of currency object.
